Let US stocks and crypto join the watchlist with full chart analysis
Stock gains an assetType (nil stays Korean for saved-data compat), and
the global market rows get star toggles plus navigation into the same
stock detail screen. Quote, FullQuote, ChartCandle, and BatchQuoteItem
move to Double prices so fractional US/crypto values decode, with
formatters showing decimals only when needed. Chart, analysis, and chart
lab route by asset kind: Yahoo candles feed /api/global/analyze/:symbol
(same engine as domestic stocks) and /api/chartlab/:code?type=us|crypto,
while the watchlist fetches global quotes per row. AI prediction and
sector peers stay Korea-only.

+++ AlphaTradingIOS/AlphaTrading/Models/GlobalQuote.swift
@@ -43,11 +43,11 @@ struct BatchQuoteItem: Decodable, Identifiable {
     var id: String { code }
     let code: String
     let name: String?
-    let price: Int?
+    let price: Double?
     let changeRate: Double?
     let changeStr: String?
     let up: Bool?
-    let volume: Int?
+    let volume: Double?
     let analysis: CandleAnalysis?
     let error: Bool?
